Add useFetchUserInfo
