+++ src/containers/header/HeaderContainer.tsx
@@ -1,12 +1,12 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useLocation, useParams } from 'react-router';
 import { push } from 'connected-react-router';
 import { Groups } from '../../reducks/groups/types';
 import { getApprovedGroups } from '../../reducks/groups/selectors';
-import { logOut, fetchUserInfo } from '../../reducks/users/operations';
+import { logOut } from '../../reducks/users/operations';
 import Header from '../../components/header/Header';
-import axios from 'axios';
+import { useFetchUserInfo } from '../../hooks/auth/useFetchUserInfo';
 
 const HeaderContainer = () => {
   const dispatch = useDispatch();
@@ -15,26 +15,7 @@ const HeaderContainer = () => {
   const pathName = useLocation().pathname.split('/')[1];
   const approvedGroups: Groups = useSelector(getApprovedGroups);
 
-  useEffect(() => {
-    const signal = axios.CancelToken.source();
-
-    const fetchUser = async () => {
-      try {
-        await dispatch(fetchUserInfo(signal));
-      } catch (error) {
-        if (error.response.status === 401) {
-          alert(error.response.data.error.message.toString());
-          dispatch(push('/login'));
-        }
-      }
-    };
-
-    fetchUser();
-
-    return () => {
-      signal.cancel();
-    };
-  }, []);
+  useFetchUserInfo(); // ユーザー情報の取得と未ログイン状態でのリダイレクト処理を行う
 
   const existsGroupWhenRouting = (path: string) => {
     if (pathName !== 'group') {
